perf!: the cross_encoding function now returns a 3D numpy array instead of a list of 2D numpy arrays.
BREAKING CHANGE: This change improves encoding speed and simplifies the code. However, this change break the pad_length option behavior for the onehot encoding.
"pad_length = 0" now returns (trimmed) encodings of length 0 instead of returning non padded encodings. Thus, the default option is now to pad the encodings to the longest sequence "pad_length= -2".
An option to return a collection of unpadded encodings will be re-introduced in the future.